refactor: `tools.schemapi.vega_expr.py` -> `tools.vega_expr.py`
https://github.com/vega/altair/pull/3600#issuecomment-2408521893
