+++ tools/schemapi/__init__.py
@@ -9,7 +9,7 @@ from tools.schemapi.codegen import (
 )
 from tools.schemapi.schemapi import SchemaBase, Undefined
 from tools.schemapi.utils import OneOrSeq, SchemaInfo
-from tools.schemapi.vega_expr import write_expr_module
+from tools.vega_expr import write_expr_module
 
 __all__ = [
     "CodeSnippet",
